Moved geometry-specific plotting to transforms and adapted diagnostic models accordingly
