+++ indica/converters/line_of_sight.py
@@ -15,6 +15,7 @@ from xarray import zeros_like
 
 from .abstractconverter_rho import Coordinates
 from .abstractconverter_rho import CoordinateTransform
+from .abstractconverter_rho import find_wall_intersections
 from ..numpy_typing import LabeledArray
 from ..numpy_typing import OnlyArray
 
@@ -73,7 +74,6 @@ class LineOfSightTransform(CoordinateTransform):
         ),
         dl: float = 0.01,
         passes: int = 1,
-        debug=False,
     ):
 
         self.name = f"{name}_line_of_sight_transform"
@@ -105,16 +105,6 @@ class LineOfSightTransform(CoordinateTransform):
         self.origin = np.array([origin_x, origin_y, origin_z]).transpose()
         self.direction = np.array([direction_x, direction_y, direction_z]).transpose()
 
-        # Wall boundaries
-        npts = 1000
-        self.angles = np.linspace(0.0, 2 * np.pi, npts)
-        self.x_wall_inner = self._machine_dims[0][0] * np.cos(self.angles)
-        self.x_wall_outer = self._machine_dims[0][1] * np.cos(self.angles)
-        self.y_wall_inner = self._machine_dims[0][0] * np.sin(self.angles)
-        self.y_wall_outer = self._machine_dims[0][1] * np.sin(self.angles)
-        self.z_wall_lower = self._machine_dims[1][0]
-        self.z_wall_upper = self._machine_dims[1][1]
-
         # Number of lines of sight
         self.x1: list = list(np.arange(0, len(origin_x)))
 
@@ -128,13 +118,9 @@ class LineOfSightTransform(CoordinateTransform):
                 direction_y[channel],
                 direction_z[channel],
             )
-            _start, _end = _find_wall_intersections(
-                origin, direction, machine_dimensions=machine_dimensions, plot=debug
+            _start, _end = find_wall_intersections(
+                origin, direction, machine_dimensions=machine_dimensions
             )
-            if debug:
-                print(f"Channel {channel}")
-                print(_start)
-                print(_end)
             x_start.append(_start[0])
             y_start.append(_start[1])
             z_start.append(_start[2])
@@ -181,7 +167,7 @@ class LineOfSightTransform(CoordinateTransform):
         x = self.x_start[x1] + (self.x_end[x1] - self.x_start[x1]) * x2
         y = self.y_start[x1] + (self.y_end[x1] - self.y_start[x1]) * x2
         z = self.z_start[x1] + (self.z_end[x1] - self.z_start[x1]) * x2
-        return np.sqrt(x**2 + y**2), z
+        return np.sqrt(x ** 2 + y ** 2), z
 
     def convert_from_Rz(
         self, R: LabeledArray, z: LabeledArray, t: LabeledArray
@@ -199,10 +185,7 @@ class LineOfSightTransform(CoordinateTransform):
             dzdx1 = 0.0
             dzdx2 = self.z_end - self.z_start
             return [
-                [
-                    2 / x * (x * dxdx1 + y * dydx1),
-                    2 / x * (x * dxdx2 + y * dydx2),
-                ],
+                [2 / x * (x * dxdx1 + y * dydx1), 2 / x * (x * dxdx2 + y * dydx2),],
                 [dzdx1, dzdx2],
             ]
 
@@ -263,11 +246,7 @@ class LineOfSightTransform(CoordinateTransform):
         return rho, theta
 
     def distance(
-        self,
-        direction: str,
-        x1: LabeledArray,
-        x2: LabeledArray,
-        t: LabeledArray,
+        self, direction: str, x1: LabeledArray, x2: LabeledArray, t: LabeledArray,
     ) -> np.ndarray:
         """Implementation of calculation of physical distances between points
         in this coordinate system. This accounts for potential toroidal skew of
@@ -284,10 +263,7 @@ class LineOfSightTransform(CoordinateTransform):
         result[{direction: slice(1, None)}] = spacings.cumsum(direction)
         return result.values
 
-    def set_dl(
-        self,
-        dl: float,
-    ) -> tuple:
+    def set_dl(self, dl: float,) -> tuple:
         """
         Set spatial resolutions of the lines of sight, and calculate spatial
         coordinates along the LOS
@@ -347,7 +323,7 @@ class LineOfSightTransform(CoordinateTransform):
         self.z = z
         self.R = R
         self.phi = phi
-        self.impact_xyz = self._impact_parameter_xyz()
+        self.impact_parameter = self.calc_impact_parameter()
 
         return x2, dl_new
 
@@ -393,11 +369,7 @@ class LineOfSightTransform(CoordinateTransform):
 
             _along_los = profile.interp(rho_poloidal=rho)
             if limit_to_sep:
-                _along_los = xr.where(
-                    rho <= 1,
-                    _along_los,
-                    0,
-                )
+                _along_los = xr.where(rho <= 1, _along_los, 0,)
             along_los.append(_along_los)
 
         self.along_los = along_los
@@ -451,10 +423,7 @@ class LineOfSightTransform(CoordinateTransform):
         Line of sight integral along the LOS
         """
         along_los = self.map_to_los(
-            profile_1d,
-            t=t,
-            limit_to_sep=limit_to_sep,
-            calc_rho=calc_rho,
+            profile_1d, t=t, limit_to_sep=limit_to_sep, calc_rho=calc_rho,
         )
 
         _los_integral = []
@@ -475,167 +444,110 @@ class LineOfSightTransform(CoordinateTransform):
 
         return los_integral
 
-    def _impact_parameter_xyz(self):
+    def calc_impact_parameter(self):
         """Calculate the impact parameter in Cartesian space"""
-        value = []
+        impact = []
         index = []
+        x = []
+        y = []
+        z = []
+        R = []
         for ch in self.x1:
             distance = np.sqrt(self.x[ch] ** 2 + self.y[ch] ** 2 + self.z[ch] ** 2)
             _index = distance.argmin()
             index.append(_index)
-            value.append(distance[_index])
+            impact.append(distance[_index])
+            x.append(self.x[ch][_index])
+            y.append(self.y[ch][_index])
+            z.append(self.z[ch][_index])
+            R.append(np.sqrt(self.x[ch][_index] ** 2 + self.y[ch][_index] ** 2))
 
-        index = xr.concat(index, "channel")
-        value = xr.concat(value, "channel")
+        impact = Dataset(
+            {
+                "index": xr.concat(index, "channel"),
+                "value": xr.concat(impact, "channel"),
+                "x": xr.concat(x, "channel"),
+                "y": xr.concat(y, "channel"),
+                "z": xr.concat(z, "channel"),
+                "R": xr.concat(R, "channel"),
+            }
+        )
 
-        impact_xyz = Dataset({"index": index, "value": value})
+        return impact
 
-        return impact_xyz
-
-    def get_plasma_boundaries(self, tplot: float = None):
-        boundaries = None
-        if hasattr(self, "equilibrium"):
-            if tplot is None:
-                if self.rho is not None:
-                    _rho = self.rho[0]
-                else:
-                    _rho = self.equilibrium.rho
-                tplot = _rho.t.sel(t=_rho.t.mean(), method="nearest").values
-
-            if self.rho is None:
-                self.convert_to_rho(t=tplot)
-
-            rho_equil = self.equilibrium.rho.sel(t=tplot, method="nearest")
-            rho_equil = xr.where(rho_equil < 1.05, rho_equil, np.nan)
-            core_ind = np.where(np.isfinite(rho_equil.interp(z=0)))[0]
-            R_lfs = rho_equil.R[core_ind[0]].values
-            R_hfs = rho_equil.R[core_ind[-1]].values
-            x_plasma_inner = R_hfs * np.cos(self.angles)
-            x_plasma_outer = R_lfs * np.cos(self.angles)
-            y_plasma_inner = R_hfs * np.sin(self.angles)
-            y_plasma_outer = R_lfs * np.sin(self.angles)
-
-            boundaries = (
-                x_plasma_inner,
-                x_plasma_outer,
-                y_plasma_inner,
-                y_plasma_outer,
-            )
-
-        return boundaries
-
-    def plot_los(self, tplot: float = None, orientation: str = "xy"):
+    def plot_los(self, tplot: float = None, orientation: str = "xy", plot_all=False):
         channels = self.x1
         cols = cm.gnuplot2(np.linspace(0.75, 0.1, len(channels), dtype=float))
 
+        wall_bounds, angles = self.get_machine_boundaries(
+            machine_dimensions=self._machine_dims
+        )
         if hasattr(self, "equilibrium"):
-            if tplot is None:
-                if self.rho is not None:
-                    _rho = self.rho[0]
-                else:
-                    _rho = self.equilibrium.rho
-                tplot = _rho.t.sel(t=_rho.t.mean(), method="nearest").values
-
-            if self.rho is None:
-                self.convert_to_rho(t=tplot)
-
-            rho_equil = self.equilibrium.rho.sel(t=tplot, method="nearest")
-            rho_equil = xr.where(rho_equil < 1.05, rho_equil, np.nan)
-            R_lfs = self.equilibrium.rmjo.sel(
-                rho_poloidal=1, t=tplot, method="nearest"
-            ).values
-            R_hfs = self.equilibrium.rmji.sel(
-                rho_poloidal=1, t=tplot, method="nearest"
-            ).values
-            x_plasma_inner = R_hfs * np.cos(self.angles)
-            x_plasma_outer = R_lfs * np.cos(self.angles)
-            y_plasma_inner = R_hfs * np.sin(self.angles)
-            y_plasma_outer = R_lfs * np.sin(self.angles)
+            equil_bounds, angles, rho_equil = self.get_equilibrium_boundaries(tplot)
             x_ax = self.equilibrium.rmag.sel(t=tplot, method="nearest").values * np.cos(
-                self.angles
+                angles
             )
             y_ax = self.equilibrium.rmag.sel(t=tplot, method="nearest").values * np.sin(
-                self.angles
+                angles
             )
 
-        if orientation is not None and orientation == "xy":
+        if orientation == "xy" or plot_all:
             plt.figure()
-            npts = len(self.x_wall_inner)
-            plt.plot(self.x_wall_inner, self.y_wall_inner, color="k")
-            plt.plot(self.x_wall_outer, self.y_wall_outer, color="k")
+            plt.plot(wall_bounds["x_in"], wall_bounds["y_in"], color="k")
+            plt.plot(wall_bounds["x_out"], wall_bounds["y_out"], color="k")
             if hasattr(self, "equilibrium"):
-                plt.plot(x_plasma_inner, y_plasma_inner, color="red")
-                plt.plot(x_plasma_outer, y_plasma_outer, color="red")
+                plt.plot(equil_bounds["x_in"], equil_bounds["y_in"], color="red")
+                plt.plot(equil_bounds["x_out"], equil_bounds["y_out"], color="red")
                 plt.plot(x_ax, y_ax, color="red", linestyle="dashed")
             for ch in self.x1:
                 plt.plot(self.x[ch], self.y[ch], color=cols[ch], linewidth=2)
+                plt.plot(
+                    self.impact_parameter["x"][ch],
+                    self.impact_parameter["y"][ch],
+                    color=cols[ch],
+                    marker="o",
+                )
             plt.xlabel("x")
             plt.ylabel("y")
             plt.axis("scaled")
 
-        if orientation is not None and orientation == "xz":
-            plt.figure()
-            plt.plot(self.x_wall_outer, [self.z_wall_upper] * npts, color="k")
-            plt.plot(self.x_wall_outer, [self.z_wall_lower] * npts, color="k")
-            plt.plot(self.x_wall_inner, [self.z_wall_upper] * npts, "w")
-            plt.plot(self.x_wall_inner, [self.z_wall_lower] * npts, "w")
-            plt.plot(
-                [self.x_wall_outer.max()] * 2,
-                [self.z_wall_lower, self.z_wall_upper],
-                color="k",
-            )
-            plt.plot(
-                [self.x_wall_inner.max()] * 2,
-                [self.z_wall_lower, self.z_wall_upper],
-                color="k",
-            )
-            plt.plot(
-                [self.x_wall_outer.min()] * 2,
-                [self.z_wall_lower, self.z_wall_upper],
-                color="k",
-            )
-            plt.plot(
-                [self.x_wall_inner.min()] * 2,
-                [self.z_wall_lower, self.z_wall_upper],
-                color="k",
-            )
-            for ch in self.x1:
-                plt.plot(self.x[ch], self.z[ch], color=cols[ch], linewidth=2)
-            plt.xlabel("x")
-            plt.ylabel("z")
-            plt.axis("scaled")
-
-        if orientation is not None and orientation == "Rz":
+        if orientation == "Rz" or plot_all:
             plt.figure()
             plt.plot(
-                [self.x_wall_outer.max()] * 2,
-                [self.z_wall_lower, self.z_wall_upper],
+                [wall_bounds["x_out"].max()] * 2,
+                [wall_bounds["z_low"], wall_bounds["z_up"]],
                 color="k",
             )
             plt.plot(
-                [self.x_wall_inner.max()] * 2,
-                [self.z_wall_lower, self.z_wall_upper],
+                [wall_bounds["x_in"].max()] * 2,
+                [wall_bounds["z_low"], wall_bounds["z_up"]],
                 color="k",
             )
             plt.plot(
-                [self.x_wall_inner.max(), self.x_wall_outer.max()],
-                [self.z_wall_lower] * 2,
+                [wall_bounds["x_in"].max(), wall_bounds["x_out"].max()],
+                [wall_bounds["z_low"]] * 2,
                 color="k",
             )
             plt.plot(
-                [self.x_wall_inner.max(), self.x_wall_outer.max()],
-                [self.z_wall_upper] * 2,
+                [wall_bounds["x_in"].max(), wall_bounds["x_out"].max()],
+                [wall_bounds["z_up"]] * 2,
                 color="k",
             )
             if hasattr(self, "equilibrium"):
                 rho_equil.plot.contour(levels=[0.01, 0.1, 0.3, 0.5, 0.7, 0.9, 0.99])
             for ch in self.x1:
                 plt.plot(self.R[ch], self.z[ch], color=cols[ch], linewidth=2)
+                plt.plot(
+                    self.impact_parameter["R"][ch],
+                    self.impact_parameter["z"][ch],
+                    color=cols[ch],
+                    marker="o",
+                )
             plt.xlabel("R")
             plt.ylabel("z")
             plt.axis("scaled")
 
-        if hasattr(self, "equilibrium") and orientation is None:
+        if hasattr(self, "equilibrium") and plot_all:
             plt.figure()
             for ch in self.x1:
                 self.rho[ch].sel(t=tplot, method="nearest").plot(
@@ -643,240 +555,3 @@ class LineOfSightTransform(CoordinateTransform):
                 )
             plt.xlabel("Path along LOS")
             plt.ylabel("Rho")
-
-
-def _find_wall_intersections(
-    origin: Tuple,
-    direction: Tuple,
-    machine_dimensions: Tuple[Tuple[float, float], Tuple[float, float]] = (
-        (1.83, 3.9),
-        (-1.75, 2.0),
-    ),
-    plot=False,
-):
-    """Function for calculating "start" and "end" positions of the line-of-sight
-    given the machine dimensions.
-
-    The end coordinate is calculated by finding the intersections with the
-    machine dimensions. If the intersection hits the inner column, the first
-    intersection point becomes the "end" position of the line-of-sight. If the
-    intersection misses the inner column, the last intersection point with the
-    outer column becomes the "end" position.
-
-    Parameters
-    ----------
-    origin
-        A Tuple (1x3) giving the X, Y and Z origin positions of the line-of-sight
-    direction
-        A Tuple (1x3) giving the X, Y and Z direction of the line-of-sight
-    machine_dimensions
-        A tuple giving the boundaries of the Tokamak in x-z space:
-        ``((xmin, xmax), (zmin, zmax)``. Defaults to values for JET.
-
-    Returns
-    -------
-    start_coordinates
-        A Tuple (1x3) giving the X, Y and Z start positions of the line-of-sight
-    end_coordinates
-        A Tuple (1x3) giving the X, Y and Z end positions of the line-of-sight
-    """
-
-    def line(_start, _end):
-        return np.linspace(_start, _end, npts)
-
-    def extremes(_start, _end):
-        return np.array(
-            [_start, _end],
-            dtype=float,
-        )
-
-    # Define XYZ lines for inner and outer walls
-    npts = 1000
-    angles = np.linspace(0.0, 2 * np.pi, npts)
-    x_wall_inner = machine_dimensions[0][0] * np.cos(angles)
-    x_wall_outer = machine_dimensions[0][1] * np.cos(angles)
-    y_wall_inner = machine_dimensions[0][0] * np.sin(angles)
-    y_wall_outer = machine_dimensions[0][1] * np.sin(angles)
-    z_wall_lower = machine_dimensions[1][0]
-    z_wall_upper = machine_dimensions[1][1]
-
-    # Define XYZ lines for LOS from origin and direction vectors
-    length = (
-        np.ceil(
-            np.max(
-                [
-                    machine_dimensions[0][1] * 2,
-                    machine_dimensions[1][1] - machine_dimensions[1][0],
-                ]
-            )
-        )
-        * 5
-    )
-    x_start = origin[0]
-    x_end = origin[0] + length * direction[0]
-    y_start = origin[1]
-    y_end = origin[1] + length * direction[1]
-    z_start = origin[2]
-    z_end = origin[2] + length * direction[2]
-
-    # Find intersections in R, z plane
-    x_line = line(x_start, x_end)
-    y_line = line(y_start, y_end)
-    z_line = line(z_start, z_end)
-    R_line = np.sqrt(x_line**2 + y_line**2)
-    indices = np.where(
-        (R_line >= machine_dimensions[0][0])
-        * (R_line <= machine_dimensions[0][1])
-        * (z_line >= machine_dimensions[1][0])
-        * (z_line <= machine_dimensions[1][1])
-    )[0]
-    if len(indices) > 0:
-        x_start = x_line[indices][0]
-        x_end = x_line[indices][-1]
-        y_start = y_line[indices][0]
-        y_end = y_line[indices][-1]
-        z_start = z_line[indices][0]
-        z_end = z_line[indices][-1]
-
-    # Find intersections with inner wall
-    xx, yx, ix, jx = intersection(
-        extremes(x_start, x_end), extremes(y_start, y_end), x_wall_inner, y_wall_inner
-    )
-    if len(xx) > 0:
-        x_line = line(x_start, x_end)
-        y_line = line(y_start, y_end)
-        z_line = line(z_start, z_end)
-        index = []
-        for i in range(len(xx)):
-            index.append(
-                np.argmin(np.sqrt((xx[i] - x_line) ** 2 + (yx[i] - y_line) ** 2))
-            )
-        indices = np.arange(np.min(index))
-        x_start = x_line[indices][0]
-        x_end = x_line[indices][-1]
-        y_start = y_line[indices][0]
-        y_end = y_line[indices][-1]
-        z_start = z_line[indices][0]
-        z_end = z_line[indices][-1]
-
-    if plot:
-        x_line = line(x_start, x_end)
-        y_line = line(y_start, y_end)
-        z_line = line(z_start, z_end)
-        R_line = np.sqrt(x_line**2 + y_line**2)
-
-        plt.figure()
-        plt.plot(x_wall_inner, y_wall_inner, color="orange")
-        plt.plot(x_wall_outer, y_wall_outer, color="orange")
-        plt.plot(x_line, y_line, "k", linewidth=2)
-        plt.xlabel("x")
-        plt.ylabel("y")
-
-        plt.figure()
-        plt.plot(x_wall_outer, [z_wall_upper] * npts, color="orange")
-        plt.plot(x_wall_outer, [z_wall_lower] * npts, color="orange")
-        plt.plot(x_wall_inner, [z_wall_upper] * npts, "w")
-        plt.plot(x_wall_inner, [z_wall_lower] * npts, "w")
-        plt.plot([x_wall_outer.max()] * 2, [z_wall_lower, z_wall_upper], color="orange")
-        plt.plot([x_wall_inner.max()] * 2, [z_wall_lower, z_wall_upper], color="orange")
-        plt.plot([x_wall_outer.min()] * 2, [z_wall_lower, z_wall_upper], color="orange")
-        plt.plot([x_wall_inner.min()] * 2, [z_wall_lower, z_wall_upper], color="orange")
-        plt.plot(x_line, z_line, "k", linewidth=2)
-        plt.xlabel("x")
-        plt.ylabel("z")
-
-        plt.figure()
-        plt.plot([x_wall_outer.max()] * 2, [z_wall_lower, z_wall_upper], color="orange")
-        plt.plot([x_wall_inner.max()] * 2, [z_wall_lower, z_wall_upper], color="orange")
-        plt.plot(
-            [x_wall_inner.max(), x_wall_outer.max()], [z_wall_lower] * 2, color="orange"
-        )
-        plt.plot(
-            [x_wall_inner.max(), x_wall_outer.max()], [z_wall_upper] * 2, color="orange"
-        )
-        plt.plot(R_line, z_line, "k", linewidth=2)
-        plt.xlabel("R")
-        plt.ylabel("z")
-
-    return (x_start, y_start, z_start), (x_end, y_end, z_end)
-
-
-def _rect_inter_inner(x1, x2):
-    n1 = x1.shape[0] - 1
-    n2 = x2.shape[0] - 1
-    X1 = np.c_[x1[:-1], x1[1:]]
-    X2 = np.c_[x2[:-1], x2[1:]]
-    S1 = np.tile(X1.min(axis=1), (n2, 1)).T
-    S2 = np.tile(X2.max(axis=1), (n1, 1))
-    S3 = np.tile(X1.max(axis=1), (n2, 1)).T
-    S4 = np.tile(X2.min(axis=1), (n1, 1))
-    return S1, S2, S3, S4
-
-
-def _rectangle_intersection_(x1, y1, x2, y2):
-    S1, S2, S3, S4 = _rect_inter_inner(x1, x2)
-    S5, S6, S7, S8 = _rect_inter_inner(y1, y2)
-
-    C1 = np.less_equal(S1, S2)
-    C2 = np.greater_equal(S3, S4)
-    C3 = np.less_equal(S5, S6)
-    C4 = np.greater_equal(S7, S8)
-
-    ii, jj = np.nonzero(C1 & C2 & C3 & C4)
-    return ii, jj
-
-
-def intersection(x1, y1, x2, y2):
-    """
-    INTERSECTIONS Intersections of curves.
-       Computes the (x,y) locations where two curves intersect.  The curves
-       can be broken with NaNs or have vertical segments.
-    usage:
-    x,y=intersection(x1,y1,x2,y2)
-        Example:
-        a, b = 1, 2
-        phi = np.linspace(3, 10, 100)
-        x1 = a*phi - b*np.sin(phi)
-        y1 = a - b*np.cos(phi)
-        x2=phi
-        y2=np.sin(phi)+2
-        x,y, ix, iy=intersection(x1,y1,x2,y2)
-        plt.plot(x1,y1,c='r')
-        plt.plot(x2,y2,c='g')
-        plt.plot(x,y,'*k')
-        plt.show()
-    """
-    ii, jj = _rectangle_intersection_(x1, y1, x2, y2)
-    n = len(ii)
-
-    dxy1 = np.diff(np.c_[x1, y1], axis=0)
-    dxy2 = np.diff(np.c_[x2, y2], axis=0)
-
-    T = np.zeros((4, n))
-    AA = np.zeros((4, 4, n))
-    AA[0:2, 2, :] = -1
-    AA[2:4, 3, :] = -1
-    AA[0::2, 0, :] = dxy1[ii, :].T
-    AA[1::2, 1, :] = dxy2[jj, :].T
-
-    BB = np.zeros((4, n))
-    BB[0, :] = -x1[ii].ravel()
-    BB[1, :] = -x2[jj].ravel()
-    BB[2, :] = -y1[ii].ravel()
-    BB[3, :] = -y2[jj].ravel()
-
-    for i in range(n):
-        try:
-            T[:, i] = np.linalg.solve(AA[:, :, i], BB[:, i])
-        except np.linalg.LinAlgError:
-            T[:, i] = np.NaN
-
-    in_range = (T[0, :] >= 0) & (T[1, :] >= 0) & (T[0, :] <= 1) & (T[1, :] <= 1)
-
-    xy0 = T[2:, in_range]
-    xy0 = xy0.T
-
-    indii = ii[in_range] + T[0, in_range]
-    indjj = jj[in_range] + T[1, in_range]
-
-    return xy0[:, 0], xy0[:, 1], indii, indjj
