+++ indica/converters/transect_rho.py
@@ -8,6 +8,8 @@ import xarray as xr
 from xarray import DataArray
 from xarray import Dataset
 from xarray import Variable
+from matplotlib import cm
+import matplotlib.pylab as plt
 
 from .abstractconverter_rho import CoordinateTransform
 from ..numpy_typing import Coordinates
@@ -74,7 +76,7 @@ class TransectCoordinates(CoordinateTransform):
         # TODO: add intersection with first walls to restrict possible coordinates
         self._machine_dims = machine_dimensions
 
-        R_positions = np.sqrt(x_positions**2 + y_positions**2)
+        R_positions = np.sqrt(x_positions ** 2 + y_positions ** 2)
         self.x_interp = interp1d(
             self.x1, x_positions, copy=False, fill_value="extrapolate"
         )
@@ -95,10 +97,7 @@ class TransectCoordinates(CoordinateTransform):
             z_positions, self.x1, copy=False, fill_value="extrapolate"
         )
         self.invert_R = interp1d(
-            R_positions,
-            self.x1,
-            copy=False,
-            fill_value="extrapolate",
+            R_positions, self.x1, copy=False, fill_value="extrapolate",
         )
 
         x, y = self.convert_to_xy(self.x1, self.x2, None)
@@ -252,11 +251,7 @@ class TransectCoordinates(CoordinateTransform):
 
         value_at_channels = profile_1d.interp(rho_poloidal=rho)
         if limit_to_sep:
-            value_at_channels = xr.where(
-                rho <= 1,
-                value_at_channels,
-                0,
-            )
+            value_at_channels = xr.where(rho <= 1, value_at_channels, 0,)
 
         self.value_at_channels = value_at_channels
 
@@ -289,3 +284,80 @@ class TransectCoordinates(CoordinateTransform):
         if not isinstance(other, self.__class__):
             return False
         return self._abstract_equals(other)
+
+    def plot_los(self, tplot: float = None, orientation: str = "xy", plot_all=False):
+        channels = self.x1
+        cols = cm.gnuplot2(np.linspace(0.75, 0.1, len(channels), dtype=float))
+
+        wall_bounds, angles = self.get_machine_boundaries(
+            machine_dimensions=self._machine_dims
+        )
+        if hasattr(self, "equilibrium"):
+            equil_bounds, angles, rho_equil = self.get_equilibrium_boundaries(tplot)
+            x_ax = self.equilibrium.rmag.sel(t=tplot, method="nearest").values * np.cos(
+                angles
+            )
+            y_ax = self.equilibrium.rmag.sel(t=tplot, method="nearest").values * np.sin(
+                angles
+            )
+
+        if orientation == "xy" or plot_all:
+            plt.figure()
+            plt.plot(wall_bounds["x_in"], wall_bounds["y_in"], color="k")
+            plt.plot(wall_bounds["x_out"], wall_bounds["y_out"], color="k")
+            if hasattr(self, "equilibrium"):
+                plt.plot(equil_bounds["x_in"], equil_bounds["y_in"], color="red")
+                plt.plot(equil_bounds["x_out"], equil_bounds["y_out"], color="red")
+                plt.plot(x_ax, y_ax, color="red", linestyle="dashed")
+            for ch in self.x1:
+                plt.scatter(self.x[ch], self.y[ch], color=cols[ch], marker="o")
+            plt.xlabel("x")
+            plt.ylabel("y")
+            plt.axis("scaled")
+
+        if orientation == "Rz" or plot_all:
+            plt.figure()
+            plt.plot(
+                [wall_bounds["x_out"].max()] * 2,
+                [wall_bounds["z_low"], wall_bounds["z_up"]],
+                color="k",
+            )
+            plt.plot(
+                [wall_bounds["x_in"].max()] * 2,
+                [wall_bounds["z_low"], wall_bounds["z_up"]],
+                color="k",
+            )
+            plt.plot(
+                [wall_bounds["x_in"].max(), wall_bounds["x_out"].max()],
+                [wall_bounds["z_low"]] * 2,
+                color="k",
+            )
+            plt.plot(
+                [wall_bounds["x_in"].max(), wall_bounds["x_out"].max()],
+                [wall_bounds["z_up"]] * 2,
+                color="k",
+            )
+            if hasattr(self, "equilibrium"):
+                rho_equil.plot.contour(levels=[0.01, 0.1, 0.3, 0.5, 0.7, 0.9, 0.99])
+            for ch in self.x1:
+                plt.scatter(self.R[ch], self.z[ch], color=cols[ch], marker="o")
+            plt.xlabel("R")
+            plt.ylabel("z")
+            plt.axis("scaled")
+
+        if hasattr(self, "equilibrium") and plot_all:
+            plt.figure()
+            plt.plot(
+                self.rho.channel,
+                self.rho.sel(t=tplot, method="nearest"),
+                color="k",
+            )
+            for ch in self.x1:
+                plt.plot(
+                    self.rho.channel[ch],
+                    self.rho.sel(channel=ch, t=tplot, method="nearest"),
+                    color=cols[ch],
+                    marker="o",
+                )
+            plt.xlabel("Channel")
+            plt.ylabel("Rho")
